+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## master (unreleased)
 
 - Serialize to JSON `cursor` value for background data migrations
+
+    Note: Run `bin/rails generate online_migrations:upgrade` if using background migrations.
+
 - Do not automatically retry stuck background migrations
 
 ## 0.35.0 (2026-07-01)